Add new implementation for Ping
Started by moving the Heartbeat handler and the rest of the
message processing to be done inside the Threading Channel.
This required the initial implementation to change because it
was done a few years ago and the implementation was very naive.

Part of this commit I also added the begining of the State class.
This object will eventually hold all the state in the library and
will only be modified inside the threading channel to make sure we
don't have any racing conditions.