Simplify and improve the code flow
Continued to remove internal update mechanisms so the library depends
more on the same update mechanisms as the user. The goal is to simplify
the internals so that it's easier to add new features

Reduced the failing tests to 34 from 200.

I've started moving some of the state to the state object but more work
is needed there.