Move all connection and connectionManager state to a single place
It started as a small refactoring but quickly spread through the whole
connection and connection manager. The Realtime channel code stayed
roughly the same but everything else changed.

The Connection State classes became much simpler. Mainly managing the
timers and how messages are processed. This way I could move all the
state management to the RealtimeWorkflow and remove all the threading
specific code. Now it's much easier to write to the state and read from
it. No locks are required.

There are still about 60 failing tests. 20 are unit tests for the
Connection state classes because they do much less and the rest are
Sandbox tests. For each one of them I need to figure out if the test is
wrong or the code is missing something.