Fix tests after refactoring
Because the logic moved from the state classes to a central place
(RealtimeWorkflow) it meant that a lot of the tests were testing the
wrong things. Some got simplified, others were refactored.

I've also added a public method to the library called GetCurrentState
which returns the full internal state and it helps when debugging
issues.