Add Fake build process
This is the initial commit for Fake. The build pipeline is getting more
complex and the main issue is we have moved CI servers a few times and
for each one we need to write a different process. Having a central
build pipeline that can be ran locally will simplify things and make
testing the build process much quicker