Fix half of the Stylecop errors. Still WIP
I started to address the style cop issues. First I made a change to the
projects to report them as errors and break the build. This made it
easier to identify. Along the way I'm adding doccumentation to all
public classes, interfaces and properties.

There is still more work to do on this.