Document and restructure Projects and solutions
The first part of this commit is to document all public interfaces,
methods and properties in the whole library. It's what took most of the
time.

The second part is to restructure the main solution and split it into a
NETFramework and NetStandard solution which can be built and tested
separately.

The third part is to add a build system which will simplify the build
process. It's using `fake` and this way we will get the build and test
process tested much easier.

The last part of the whole restructure is still remaining. I need to
finish implementing Disposable properly, finish all the todos and setup
the build and test process to be run in AzureDevOps.