Remove handled TODO comments
When processing a send message we don't need to worry if the transport is closed any more. When trying to send the message it will throw an exception which we catch and generate a TransportClose command anyway. Recently we implemented a way to handle that edge case and we discard the extra TransportClose.