Implement RTP17c2 spec item for presence
Moved away from the presence listening to Channel events so it's easier
to follow the flow. We had a failing test RTP5c2 which started failing
after I updated the call to channel.EmitUpdate.
Now the channel AttachSerial is set by the ChannelMessageProcessor when
an Attached message is received and channel.SetState and the
ChannelMessageProcess call into the Presence object to notify it when
something of relevance happens.