Refactor for handling connection fallback states
This refactor includes both the realtime and rest clients.
Realtime - I moved the connecting error logic to the Realtime workflow and created a special command to handle the different between receiving a disconnected and an error message from the server while connecting.
Rest:
The logic now follows the spec so we retry all possible options and on every request and save a preferred host when we succeed