Implement Delta recovery when decoding or a precondition fails
First I made the DecodeRecovery property read the value in a thread safe way because I realised it could be accessed from multiple places.
I also add the ErrorInfo to the channel state transition to Attaching during delta recovery process.
Finally I updated the ChannelMessageProcessor to first validate the DeltaExtras From property to ensure we are processing the correct message and also added a check to make sure messages are not processed while the channel is in Attaching state.