Update tests for RTL7d and RTL7e
When I fixed the code to stop emitting the error the above 2 tests started failing as well. I updated the test for RTL7d to stop looking for the error.
I also found the test for RTL7e needed updating as well. It was not checking the residual encoding. I updated the test to check there is an error message and also the correct encoding is kept