Update PendingMessages to be a List instead of a Queue
Even thought a queue is better suited for the job if we don't use it we fix issue #384.
I originally started by copying the source code of the Queue<> class from the .net core library but after I found we only use a Queues in 1 place it was easier to switch to using a List.
ConcurrentQueue