Fix auth token clock-skew (#439)
* Removed token expiry check from auth flow

* Update Token Expiry checking to match RSA4b1

* Rename ServerTimeOffset to ServerTime. The fact that it is a datetimeoffset doesn't add any value to the code
* Change the logic for IsValidToken to only validate the expiry if there is serverTime and if we are dealing with a TokenDetails object with an Expiry
* Add extra logging around getting and setting the server time

TODO: We still don't check whether the local time has drifted

* Update Workflow to make sure Reauthentication happens

Previously this functionality was deep within AblyAuth but when we started checking server time I had to move it out.

* Fix typo and correct logic for RTN15h because the test was actually testing RTN15h2

* Add tests for RSA4b1

(RSA4b1) Client libraries can optionally save a round-trip request to the Ably service for expired tokens by detecting when a token has expired when all of the following applies: the current token is a TokenDetails object with an expires attribute; the library has previously queried the time from the Ably service and persisted the local clock offset according to RSA10k; the expires time has passed based on the Ably service time and not the local clock (which is not guaranteed to be accurate)

* Fix test that depends on Server time

* Remove unnecessary test

Nowhere in the spec do we go straight to failed if we get an exception when trying to renew a token

* Fix test for MessageExtras

* Remove TokenDetails Expire method

It was used only in one place that was the internal AblyAuth class. It went really easily without breaking any tests

* Fix Idempotent test

We were not passing a messageId so it's no wonder it failed a lot. It was a fluke if it passed

* Improve ResumeRequest_ConnectedProtocolMessageWithSameConnectionId_WithNoError test

Now it passes every time

* Remove debug log level

* Mark test as intermittent

I think there is a logic issue which breaks at specific edge cases. I haven't figured a good way to recreate it so for now I'm marking it to come back to it later

* Code style fix

Co-authored-by: Martin Georgiev <marto83@gmail.com>