Update SendMessage command to handle failed message send
There was an edge case where a message was just sent to the transport and the socket was already closed. We didn't have a way to let the library know the message wasn't actually sent. In these cases the message was not retried correctly.
Now if _socket has been disposed and we call send message before we have reacted to the socket being disposed we will queue the message to be sent later if possible. This fixed several intermittent tests around making sure the Detach or Attach message is resent if the connecting drops.
A had to make a breaking change to the ITransport interface which should be documented in the release notes.