Fix skipped test 'TaskUtilsTests.FailingTaskShouldHaveExceptionHandled'
This corresponds to [Issue 829](https://github.com/ably/ably-dotnet/issues/829)

The fundamental problem is that the test can legitimately return `false` if it
takes longer to execute than the supplied timeout of 500-ms.  Added a `TimeoutExpired`
boolean property that allows the test to detect this and check for the correct value
depending on whether the test completed in time, or the timeout expired.