Be explicit about ignored return values
By being explicit we communicate to future developers that we actually intended to ignore the return value, we also inform the compiler which then won't raise a warning. An added advantage is that the ignored return value is more obvious at the point of call, rather than possibly being a call to a method that returns `void`. Lastly the `_ =` is very grep-able.