Add message extras to the Message equality comparison
Sachin caught the issue when serialising a ProtocolMessage containing Messages with only Message extras and no other properties. As Message extras was not in the Equality comparison the `OnSerializing` method in `ProtocolMessage` thought all the messages were empty and set the whole `Messages` property to `null` thus causing issues.
I don't think it would have caused any real user issues because having a message with only Extras is a pointless. It did however cause issues when Sachin enabled Debug logging in the RealtimeWorkflow class.