Delete unnecessary recover error test
The test removed in this commit expects that using an invalid recovery
key leads to an ERROR protocol message being returned which thus fails
the connection, but this isn't part of the spec, we make no guarantee
that invalid recovery keys are rejected, and in fact returning a
CONNECTED protocol message with the error field indicating the recovery
key was invalid is a perfectly good response, which is in fact what we
will start to do in the future.

The functionality that an ERROR protocol message is treated as fatal
and fails the connection is still tested elsewhere, so rmeoving this
test doesn't materially reduce the amount of test coverage.

Signed-off-by: Lewis Marshall <lewis.marshall@ably.com>